requirements.txt
tensorflow<=1.12
numpy
imageio
pillow